move Qt-related stuff out of model and into model-networking
