+++ libraries/entities/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(TARGET_NAME entities)
 setup_hifi_library(Network Script)
-link_hifi_libraries(avatars shared audio octree model fbx networking animation)
+link_hifi_libraries(avatars shared audio octree model model-networking fbx networking animation)
 
 target_bullet()
